pyproject.toml
[project]
name = "system-trading-starterkit"
version = "0.1.0"
description = "한국 주식 시스템 트레이딩 스타터킷 (Kiwoom REST API + Python asyncio)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "apscheduler>=3.11",
    "pandas>=2.2",
    "numpy>=2.0",
    "ta>=0.11",
    "aiosqlite>=0.20",
    "python-telegram-bot>=21.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["trading_bot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
